Support installing just in mulitple operating systems.
This follows the official just instructions for [installing pre-built
binaries](https://github.com/casey/just#pre-built-binaries).

However, the current version of https://just.systems/install.sh doesn't
quite work in windows. I have fixed that and [proposed the changes
upstream](https://github.com/casey/just/pull/1474), and include a copy
of the fixed script in this repo for now.

As part of this, have added tests for this action, which runs on
multiple OSs, in order to test the cross platfoem installation

Also, allow installing a specific version of just.