Add `install-uv` option
Similar to the `install-just` option, add an `install-uv` option to
install `uv` using astral's `setup-uv` GitHub action. No default value
is set here, but the repo-template will be setting this to `true`.

Using a commit sha following the docs on
[Using third-party actions](https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions).
Dependabot should be able to recognise that the sha corresponds to a
specific version and propose updates.

`uv` is capable of installing `just`, but I haven't altered the `just`
installation part of the action. We may wish to revisit how `just` is
installed later, including using `uv` to install it if `install-uv` is
set to `true`. We can also use `setup-just`. For flexibility in the
future, I have placed the `uv` installation step before the `just`
installation step.